refactor: OrderLineItems를 도메인 클래스로 정의
